requirements.txt
# SMS Spam Detection — Python dependencies
# Install with: pip install -r requirements.txt

# Core ML
scikit-learn>=1.3
numpy>=1.24
pandas>=2.0

# NLP
nltk>=3.8

# Visualisation
matplotlib>=3.7
seaborn>=0.12

# Persistence / utilities
joblib>=1.3

# UI
streamlit>=1.32

# Testing
pytest>=7.4

# DVC — Data Version Control
dvc>=3.0
pyyaml>=6.0

# MLflow — Experiment Tracking
mlflow>=2.10

# Monitoring
prometheus-client>=0.20
